// src/interfaces/auth/IUserCredentials.ts
export interface IUserCredentials {
    userId?: string;
    userType?: string;
    userNom?: string;
    userPrenom?: string;
    userGenre?: string;
    userMail?: string;
    userPass?: string;
    newUserPass?: string;
    userCreat?: string;
    accessToken?: string;
    confirmedPassword?: string;
    accountValid?: string;
    key?: string;
    keyType?: string;
    userConnected?: boolean;
}